Fixed Context issue with Layout component
