+++ src/theme/Root.js
@@ -1,0 +1,7 @@
+import { MsalProvider } from "@azure/msal-react";
+import useMsalAuth from "../hooks/useMsalAuth";
+
+export default function Root({ children }) {
+  const { msalInstance } = useMsalAuth();
+  return <MsalProvider instance={msalInstance}>{children}</MsalProvider>;
+}
